LPS-17027 Remembering the right nested fields context
We were getting into an infinite loop when running this test:

ObjectEntryRelatedObjectsResourceTest.testPostCustomObjectEntryWithNestedSystemObjectEntry

Before lazy loading, when calling NestedFieldsSupplier.supply the depth
was being incremented, then the data retrieved (with more potential nested
calls to supply) and then decremented, wich was fine.

With lazy loading, the retrieval of data no longer happens in between the
calls to increment/decrement but later on when serializing the data.
This causes the depth to always be 0.

This fix ensures the supplier is run with the same context present when
it was created thus keeping the right depth and breaking the endless
loop.