LPD-19409 Define a class to contribute to the DSLQueryBuilder.
This class will split the sort into different ones, could be complex (relationships) or simple (fields). For each relationship type, a DSLQueryBuilder will be developed and the main DSLQueryBuilder will invoke the one that is the responsible based on the relationship type.

At the end, the contributor for the field contributes to the DSLQueryBuilder based on the latest table processed