LPD-18195 Provide all Base*ResourceImpls with a httpServletRequest which the ctx attribute, no matter where resourceImpls are created. Fix covers the 2 *ResourceImpl creation scenarios  * From the *ResourceFactoryImpl, used from DXP codebase to interact with resources. This was fixed before in the ResourceImpl.setContextHttpServletRequest setter. Now it's moved to the factory  * From the ContextContainerRequestFilter (part of JAXRS message pipeline), used when calling rest endpoints. As this uses reflection to set the field ResourceImpl.setContextHttpServletRequest setter was not being hit, hence the failure.
Together, both fixes enables resourceImpls to properly dispatch JSPs when a page render is required as part of the resource methods or the indexer transaction callbacks.
Fixes SitePageResourceImpl not being able to add widgets to the page being created if they include JSPs in portal-web servlet context.

Currently, we use SitePageResourceImpl in tests to create site pages:
 * From playwright via REST api calls
 * From the BundleSiteInitializerTest integration test