LPD-25201 Create an extensible point. With the new extension point, we will only extend it from Objects, as the Object's scopes are the only one that must be allowed ignoring case so far. For the rest of the scope aliases, we will mantain the exact values.
For a given Object Definition called Student, in the past we used to have a scope aliases like `C_Student`. However, it was necessary to update the JAX-RS application name from `C_Student` to `c_student`.

With this change we will keep supporting "old-format" scope aliases from Client Extensions, so for a scope aliases `C_Student.everything.read`, we will transform into `c_student.everyhing.read` in case it exists, otherwise the scope aliases will not be inserted into the OAuth2 application, just as before this change.