LPD-25201 Add ModelListener to update the scopeAliases by calling the map method of the ScopeAliasesMapper.
This is necessary because once the validations have passed, the initial scope aliases are the ones that are inserted into the DB, so we need to update them with the mapped values before inserting.

As the LiferayOAuth2Scope interface does not contain the scopeAlias, it is not possible to propagate the value when calling the ScopeLocator methods.