LPD-35336 Simplifying
We now use the same components used by batch engine. This means that
for objects, the _objectDefinition is already set and we do not need
to handle that. We just need to call the setCompany as batch engine
does.

I've also simplified the company id injection and we will just be
getting it from the thread local. This is because modifying the
OpenAPIResource interface is risky as it could cause breaking changes
with previous versions of REST Builder used by customers.

With this commit we reduce the scope of the change just to the vulcan
infrastructure, making it transparent to the other modules except for
objects in which we do need to set certain properties when registering
the components.