LPD-51446 Include the dependency with LICENSE_INSTALL instead. The reason:
In the MainServlet class, the method 'DBUpgrader.upgradeModules()' is invoked before 'StartupHelperUtil.setUpgrading(false)'. Inside the 'DBUpgrader.upgradeModules()' method, the 'ModuleServiceLifecycle.class' is registered in OSGi with the property 'module.service.lifecycle' set to 'portlets.initialized'. Then, there is an amount of time where the ModuleServiceLifecycle has been registered in OSGi but the 'StartupHelperUtil._upgrading' property is still true. That causes some indexing problems with the elements that are being created by Batch, as the BundleTracker is active since the ModuleServiceLifecycle was registered, and then, all the elements (ObjectDefinition, etc) are created but not indexed.

We have added this change as a workaround until the ModuleServiceLifecycle registration issue is fixed. It is still not clear which will be the final solution, could be setting the property upgrading to false before registering the OSGi component, or even registering a new ModuleServiceLifecycle with a different new property value. That will be handled by the Database Infrastructure Team

Co-authored-by: Alejandro Tardín <alejandro.tardin@liferay.com>
Co-authored-by: Alberto Chaparro <alberto.chaparro@liferay.com>